requirements-dev.txt
-r requirements.txt

#######################################################
### Debug                                           ###
#######################################################
django-debug-toolbar==1.4.0
